+++ app/javascript/vue/shared/Filter/constants/containerItem.js
@@ -1,0 +1,8 @@
+export const CONTAINER_ITEM = [
+  'id',
+  'parent_id',
+  'disposition',
+  'disposition_x',
+  'disposition_y',
+  'disposition_z'
+]