+++ app/javascript/vue/tasks/collection_objects/filter/constants/layouts.js
@@ -4,7 +4,8 @@ import {
   DWC_OCCURRENCE_PROPERTIES,
   REPOSITORY_PROPERTIES,
   TAXON_DETERMINATION_PROPERTIES,
-  IDENTIFIER_PROPERTIES
+  IDENTIFIER_PROPERTIES,
+  CONTAINER_ITEM
 } from '@/shared/Filter/constants'
 
 export const LAYOUTS = {
@@ -17,6 +18,7 @@ export const LAYOUTS = {
       taxon_determinations: TAXON_DETERMINATION_PROPERTIES,
       dwc_occurrence: DWC_OCCURRENCE_PROPERTIES,
       identifiers: IDENTIFIER_PROPERTIES,
+      container_item: CONTAINER_ITEM,
       data_attributes: { show: true }
     }
   },
@@ -36,6 +38,7 @@ export const LAYOUTS = {
       collecting_event: COLLECTING_EVENT_PROPERTIES,
       taxon_determinations: TAXON_DETERMINATION_PROPERTIES,
       identifiers: IDENTIFIER_PROPERTIES,
+      container_item: CONTAINER_ITEM,
       data_attributes: { show: true }
     }
   },